+++ scene/wlf_line_node.c
@@ -22,7 +22,7 @@ static void bounds(const struct wlf_line_shape *shape,
 }
 
 struct wlf_line_node *wlf_line_node_create(struct wlf_scene_node *parent,
-		double x, double y, struct wlf_line_shape *shape) {
+		int x, int y, struct wlf_line_shape *shape) {
 	if (shape == NULL) return NULL;
 	struct wlf_line_node *node = calloc(1, sizeof(*node));
 	if (node == NULL) return NULL;
@@ -58,7 +58,7 @@ static void render_at(struct wlf_line_node *node,
 void wlf_line_node_render(struct wlf_line_node *node,
 		struct wlf_line_pass *pass, struct wlf_render_target_info *target,
 		const pixman_region32_t *clip) {
-	double x, y;
+	int x, y;
 	if (node == NULL || !wlf_scene_node_coords(&node->base, &x, &y)) {
 		return;
 	}
