+++ scene/wlf_path_node.c
@@ -32,7 +32,7 @@ static bool bounds(const struct wlf_path_shape *shape,
 }
 
 struct wlf_path_node *wlf_path_node_create(struct wlf_scene_node *parent,
-		double x, double y, struct wlf_path_shape *shape) {
+		int x, int y, struct wlf_path_shape *shape) {
 	if (shape == NULL) return NULL;
 	struct wlf_path_node *node = calloc(1, sizeof(*node));
 	if (node == NULL) return NULL;
@@ -68,7 +68,7 @@ static void render_at(struct wlf_path_node *node,
 void wlf_path_node_render(struct wlf_path_node *node,
 		struct wlf_path_pass *pass, struct wlf_render_target_info *target,
 		const pixman_region32_t *clip) {
-	double x, y;
+	int x, y;
 	if (node == NULL || !wlf_scene_node_coords(&node->base, &x, &y)) {
 		return;
 	}
